+++ apps/boruta_gateway/lib/boruta_gateway/https_gateway.ex
@@ -15,6 +15,7 @@ defmodule BorutaGateway.HttpsGateway do
   alias BorutaGateway.Certificate
   alias BorutaGateway.HttpRequest
   alias BorutaGateway.HttpsGateway.Authorization
+  alias BorutaGateway.NoiseCancelling
   alias BorutaGateway.Upstreams
   alias BorutaGateway.Upstreams.Upstream
 
@@ -254,6 +255,7 @@ defmodule BorutaGateway.HttpsGateway do
         path_info = path_info(path)
 
         with %Upstream{} = upstream <- match_upstream(state.match_function, header, path_info),
+             :ok <- NoiseCancelling.check(upstream, method, path, state.remote_ip),
              :ok <- rate_limit(socket, upstream),
              {:ok, token} <- Authorization.authorize(header, method, upstream) do
           connect_upstream(
@@ -281,6 +283,18 @@ defmodule BorutaGateway.HttpsGateway do
             )
 
             log_exchange(state, start, request_id, method, path, nil, 404, :failure)
+
+            {:noreply, close_downstream(socket, state)}
+
+          {:noise, _prediction} ->
+            response = "No upstream has been found corresponding to the given request."
+
+            send_downstream(
+              socket,
+              "HTTP/1.1 404 Not Found\r\n" <>
+                "Content-Length: 62\r\n\r\n" <>
+                response
+            )
 
             {:noreply, close_downstream(socket, state)}